requirements.txt
streamlit
openai
google-generativeai
python-dotenv
beautifulsoup4
requests
faiss-cpu
sentence-transformers
numpy